frontend/projects/bookolog-ui-kit/src/lib/ui-button/components/icon-button/icon-button.component.ts
import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
import { ButtonComponent } from '../button/button.component';

@Component({
  selector: 'button[uiIconButton], input[uiIconButton][type=submit], a[uiIconButton]',
  templateUrl: './icon-button.component.html',
  styleUrls: ['./icon-button.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush,
  encapsulation: ViewEncapsulation.None,
})
export class IconButtonComponent extends ButtonComponent {
}
